backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
stripe>=9.9.0
supabase>=2.4.0
python-dotenv>=1.0.0
pydantic>=2.7.0
httpx>=0.27.0
